warning LNK4098: defaultlib 'LIBCMT' conflict fixed + VertexArray cast warning fixed

+++ Engine/src/Core/OpenGL/VertexArray.cpp
@@ -41,7 +41,7 @@ namespace SW {
 			glEnableVertexAttribArray(index);
 
 			glVertexAttribPointer(index, element.GetComponentCount(), ShaderDataTypeToOpenGLBaseType(element.Type),
-				GL_FALSE, layout.GetStride(), reinterpret_cast<const void*>(element.Offset));
+				GL_FALSE, layout.GetStride(), reinterpret_cast<const void*>(static_cast<uintptr_t>(element.Offset)));
 
 			index++;
 		}
